Refactored code to use cell based maze approach.
Updated the entire codebase to use a Cell[][] instead of int[][] to
represent the maze, where a Cell is an object with mutable properties.
Updated Generator.java to translate binary maze to cell maze.  Updated
display methods to use the cell based implementation.  Updated the maze
class to represent and store data differently.

Still need to write updated solver methods for the Cell based
implementation.